VestaAppServer/Django/RAG/requirements.txt
Django>=5.2,<7.0
djangorestframework>=3.15,<4.0
python-dotenv>=1.0,<2.0
chromadb>=1.4.0
httpx>=0.27,<1.0
tqdm>=4.66,<5.0
django-cors-headers>=4.0.0